+++ src/screens/App/Verification/UnderVerification/UnderVerification.module.scss
@@ -1,5 +1,5 @@
-@import "../../../styles/Colors";
-@import "../../../styles/Properties";
+@import "../../../../styles/Colors";
+@import "../../../../styles/Properties";
 
 .addressContainer {
   display: flex;
@@ -27,29 +27,6 @@
   line-height: 22px;
 }
 
-.termsContainer {
-  margin-top: 16px;
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: flex-start;
-
-  .termsCheckbox {
-    width: 20px;
-    height: 20px;
-    border-radius: 4px;
-  }
-
-  .termsText {
-    padding-top: 8px;
-    margin-left: 8px;
-  }
-
-  .termsLink {
-    color: $orange-500;
-  }
-}
-
 .buttonContainer {
   display: flex;
   flex-direction: row;
@@ -62,7 +39,7 @@
     font-size: 16px;
   }
 
-  .saveButton {
+  .resubmitButton {
     height: 48px;
     width: 176px;
     margin-left: 28px;
@@ -76,16 +53,6 @@
       h4 {
         color: $white;
       }
-    }
-  }
-
-  .submitButton {
-    height: 48px;
-    width: 176px;
-    margin-left: 28px;
-
-    h4 {
-      color: $white;
     }
   }
 }
