+++ src/screens/App/Verification/VerificationForm/Verification.module.scss
@@ -1,0 +1,91 @@
+@import "../../../../styles/Colors.scss";
+@import "../../../../styles/Properties";
+
+.addressContainer {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+
+  .city,
+  .state {
+    flex: 2;
+    margin-right: 20px;
+  }
+
+  .zip {
+    flex: 1;
+  }
+}
+
+.missionHeader {
+  margin-top: 52px;
+}
+
+.missionText {
+  margin-top: 24px;
+  margin-bottom: 16px !important;
+  line-height: 22px;
+}
+
+.termsContainer {
+  margin-top: 16px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-start;
+
+  .termsCheckbox {
+    width: 20px;
+    height: 20px;
+    border-radius: 4px;
+  }
+
+  .termsText {
+    padding-top: 8px;
+    margin-left: 8px;
+  }
+
+  .termsLink {
+    color: $orange-500;
+  }
+}
+
+.buttonContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: flex-end;
+
+  margin-top: 48px;
+
+  h4 {
+    font-size: 16px;
+  }
+
+  .saveButton {
+    height: 48px;
+    width: 176px;
+    margin-left: 28px;
+    background-color: $gray-5;
+
+    h4 {
+      color: $orange-500;
+    }
+
+    &:hover {
+      h4 {
+        color: $white;
+      }
+    }
+  }
+
+  .submitButton {
+    height: 48px;
+    width: 176px;
+    margin-left: 28px;
+
+    h4 {
+      color: $white;
+    }
+  }
+}
